requirements_dev.txt
pip==20.0.2
asyncpg==0.25.0
pyyaml==5.3.1